refactor: QuackText 위치 internal 로 이동
